Refactor the test_helper file.
