+++ lists/singly_linked_list.go
