+++ lists/slice_list.go
