+++ lists/slice_list_test.go
@@ -8,18 +8,10 @@ import (
 	adts "github.com/johnsrd7/go-adts"
 )
 
-func printSinglyLinkedList(ll *SinglyLinkedList) {
-	for tmp := ll.head; tmp != nil; tmp = tmp.next {
-		fmt.Printf("%v -> ", tmp.elt)
-	}
+func TestMakeSliceList(t *testing.T) {
+	list := MakeSliceList()
 
-	fmt.Println()
-}
-
-func TestMakeSinglyLinkedList(t *testing.T) {
-	list := MakeSinglyLinkedList()
-
-	if list.len != 0 {
+	if len(list.backer) != 0 {
 		t.Error("Length of empty list should be 0")
 	}
 	if list.threadSafe {
@@ -30,7 +22,7 @@ func TestMakeSinglyLinkedList(t *testing.T) {
 	}
 
 	var l List
-	l = MakeSinglyLinkedList()
+	l = MakeSliceList()
 
 	if l.Len() != 0 {
 		t.Error("Length of empty list should be 0.")
@@ -41,11 +33,11 @@ func TestMakeSinglyLinkedList(t *testing.T) {
 // Test Container Methods
 // -------------------------------------------------------
 
-func TestSinglyLinkedListLen(t *testing.T) {
-	list := MakeSinglyLinkedList()
+func TestSliceListLen(t *testing.T) {
+	list := MakeSliceList()
 
 	for i := 0; i < 50; i++ {
-		list.Add(intElt(i))
+		list.backer = append(list.backer, adts.IntElt(i))
 
 		if list.Len() != i+1 {
 			t.Errorf("List should have length %d, actual length: %d", i+1, list.Len())
@@ -54,24 +46,24 @@ func TestSinglyLinkedListLen(t *testing.T) {
 	}
 }
 
-func TestSinglyLinkedListIsEmpty(t *testing.T) {
-	list := MakeSinglyLinkedList()
+func TestSliceListIsEmpty(t *testing.T) {
+	list := MakeSliceList()
 
 	if !list.IsEmpty() {
 		t.Errorf("Empty list should return true for IsEmpty.\n")
 	}
 
-	list.Add(intElt(0))
+	list.Add(adts.IntElt(0))
 	if list.IsEmpty() {
 		t.Errorf("List with 1 element should not return true for IsEmpty.\n")
 	}
 }
 
-func TestSinglyLinkedListClear(t *testing.T) {
-	list := MakeSinglyLinkedList()
+func TestSliceListClear(t *testing.T) {
+	list := MakeSliceList()
 
 	for i := 0; i < 100; i++ {
-		list.Add(intElt(i))
+		list.Add(adts.IntElt(i))
 	}
 
 	if list.Len() != 100 {
@@ -84,67 +76,68 @@ func TestSinglyLinkedListClear(t *testing.T) {
 	}
 }
 
-func TestSinglyLinkedListAdd(t *testing.T) {
-	expected := []adts.ContainerElement{}
+func TestSliceListAdd(t *testing.T) {
+	vals := make(map[int][]int) // value -> slice of indices
 	r := rand.New(rand.NewSource(99))
 
-	list := MakeSinglyLinkedList()
+	list := MakeSliceList()
 
 	for i := 0; i < 1000; i++ {
-		v := intElt(r.Int())
+		v := r.Int()
 
-		expected = append(expected, v)
+		vals[v] = append(vals[v], i)
 
-		if !list.Add(v) {
+		if !list.Add(adts.IntElt(v)) {
 			t.Errorf("Failed to add %d to list\n", v)
 			return
 		}
 
-		if list.Len() != i+1 {
-			t.Errorf("List size not correct. Expected: %d, Actual: %d", i+1, list.len)
+		if len(list.backer) != i+1 {
+			t.Errorf("List size not correct. Expected: %d, Actual: %d", i+1, len(list.backer))
 			return
 		}
-
-		tmp := list.head
-		for idx, val := range expected {
-			if !tmp.elt.Equals(val) {
-				t.Errorf("Add failed to add the value to the proper index (%d) | (idx,val) - Expected: (%d, %v), Actual: (%d, %v)\n",
-					i, idx, val, idx, tmp.elt)
-				return
+		for v, idxs := range vals {
+			for _, idx := range idxs {
+				if idx >= len(list.backer) {
+					fmt.Printf("Idx: %d, Len: %d\n", idx, len(list.backer))
+					continue
+				}
+				if !list.backer[idx].Equals(adts.IntElt(v)) {
+					t.Errorf("Add failed to add the value to the proper index | (idx,val) - Expected: (%d, %d), Actual: (%d, %v)",
+						idx, v, idx, list.backer[idx])
+				}
 			}
-
-			tmp = tmp.next
 		}
 	}
 }
 
-func TestSinglyLinkedListContains(t *testing.T) {
+func TestSliceListContains(t *testing.T) {
 	vals := make(map[int]bool)
 	r := rand.New(rand.NewSource(99))
 
-	list := MakeSinglyLinkedList()
+	list := MakeSliceList()
 
 	for i := 0; i < 1000; i++ {
 		v := r.Int()
 
 		vals[v] = true
 
-		list.Add(intElt(v))
+		list.Add(adts.IntElt(v))
 
 		for v := range vals {
-			if !list.Contains(intElt(v)) {
+			if !list.Contains(adts.IntElt(v)) {
 				t.Errorf("Contains failed to find the value (%d) in the list.", v)
 			}
 		}
 	}
 }
 
-func TestSinglyLinkedListRemove(t *testing.T) {
+func TestSliceListRemove(t *testing.T) {
 	max := 1000
 
-	list := MakeSinglyLinkedList()
+	list := MakeSliceList()
 	for i := 0; i < max; i++ {
-		list.Add(intElt(i))
+		list.Add(adts.IntElt(i))
 	}
 
 	for i := 0; i < max; i++ {
@@ -156,7 +149,7 @@ func TestSinglyLinkedListRemove(t *testing.T) {
 		}
 
 		if list.Len() != max-1-i {
-			t.Errorf("(%d) Expected length: %d, Actual length: %d", i, max-1-i, list.Len())
+			t.Errorf("Expected length: %d, Actual length: %d", max-1-i, list.Len())
 			return
 		}
 
@@ -166,8 +159,13 @@ func TestSinglyLinkedListRemove(t *testing.T) {
 		}
 	}
 
-	if list.Len() != 0 {
-		t.Error("Length of list should be 0 after all elements are removed.\n")
+	// Now we want to check that the resize didn't break the ability to add.
+	for i := 0; i < max; i++ {
+		list.Add(adts.IntElt(i))
+		if list.Len() != i+1 {
+			t.Error("Unable to add after removing all elements.")
+			return
+		}
 	}
 }
 
@@ -175,38 +173,38 @@ func TestSinglyLinkedListRemove(t *testing.T) {
 // Test List Methods
 // -------------------------------------------------------
 
-func TestGet(t *testing.T) {
+func TestSliceListGet(t *testing.T) {
 	r := rand.New(rand.NewSource(99))
 
 	expected := []int{}
-	list := MakeSinglyLinkedList()
+	list := MakeSliceList()
 
 	for i := 0; i < 1000; i++ {
 		v := r.Int()
 
 		expected = append(expected, v)
-		list.Add(intElt(v))
+		list.Add(adts.IntElt(v))
 	}
 
 	for idx, v := range expected {
-		if !list.Get(idx).Equals(intElt(v)) {
+		if !list.Get(idx).Equals(adts.IntElt(v)) {
 			t.Errorf("Expected: %d, Actual: %v", v, list.Get(idx))
 			return
 		}
 	}
 }
 
-func TestSet(t *testing.T) {
-	list := MakeSinglyLinkedList()
+func TestSliceListSet(t *testing.T) {
+	list := MakeSliceList()
 
 	for i := 0; i < 100; i++ {
-		list.Add(intElt(i))
+		list.Add(adts.IntElt(i))
 	}
 
 	for i := 0; i < 100; i++ {
-		list.Set(i, intElt(i*2))
+		list.Set(i, adts.IntElt(i*2))
 
-		if !list.Get(i).Equals(intElt(i * 2)) {
+		if !list.Get(i).Equals(adts.IntElt(i * 2)) {
 			t.Errorf("Set did not set element at index %d properly. Expected: %v, Actual: %v\n", i, i*2, list.Get(i))
 			return
 		}
